requirements.txt
django-celery-results
lxml
redis